Fixed issue with one-off errors
Essentially there were two problems. First, we were not getting the
Jaccard calculation for the LAST period (e.g. comparing the second last
to the last). Second, we were not getting the Jaccard calculation if the
event data only had 2 periods (e.g. Period 45 and 60). This is fixed
now! Plus added another column to better sort the data.